bond and angle Gamess inputs generation
The ability to generate Gamess inputs for bond and angle scans is
added. The Ginp.py input needs a “bond”, “angle” or “diva” keyword at
the end of the line. If there is none, “diva” is assumed by default.
(Numerous small improvements in _GeomCalc.py, GeomStr.py, IO.py and
Ginp.py.)